+++ test/Microsoft.IdentityModel.Tokens.Saml.Tests/SamlTestData.cs
@@ -20,6 +20,7 @@ namespace Microsoft.IdentityModel.Tokens.Saml.Tests
                     TestId = "TokenExpired",
                     TokenDescriptor = new SecurityTokenDescriptor
                     {
+                        Audience = Default.Audience,
                         Subject = new ClaimsIdentity(Default.SamlClaims),
                         Expires = DateTime.UtcNow.Subtract(new TimeSpan(0, 10, 0)),
                         IssuedAt = DateTime.UtcNow.Subtract(new TimeSpan(1, 0, 0)),
@@ -39,6 +40,7 @@ namespace Microsoft.IdentityModel.Tokens.Saml.Tests
                     TestId = "InvalidIssuer",
                     TokenDescriptor = new SecurityTokenDescriptor
                     {
+                        Audience = Default.Audience,
                         Subject = new ClaimsIdentity(Default.SamlClaims),
                         SigningCredentials = Default.AsymmetricSigningCredentials,
                         Issuer = Default.Issuer,
@@ -54,6 +56,7 @@ namespace Microsoft.IdentityModel.Tokens.Saml.Tests
                     TestId = "ExpiredAndInvalidIssuer",
                     TokenDescriptor = new SecurityTokenDescriptor
                     {
+                        Audience = Default.Audience,
                         Subject = new ClaimsIdentity(Default.SamlClaims),
                         Expires = DateTime.UtcNow.Subtract(new TimeSpan(0, 10, 0)),
                         IssuedAt = DateTime.UtcNow.Subtract(new TimeSpan(1, 0, 0)),
@@ -74,6 +77,7 @@ namespace Microsoft.IdentityModel.Tokens.Saml.Tests
                     TestId = "KeysDontMatchValidLifetimeAndIssuer",
                     TokenDescriptor = new SecurityTokenDescriptor
                     {
+                        Audience = Default.Audience,
                         Subject = new ClaimsIdentity(Default.SamlClaims),
                         SigningCredentials = Default.AsymmetricSigningCredentials,
                         Issuer = Default.Issuer,